convert time capsule to TS
